+++ kgml-core/src/commonMain/kotlin/dev/karmakrafts/kgml/vector/Vector3f.kt
@@ -258,7 +258,7 @@ data class Vector3f( // @formatter:off
      * @param other The other vector.
      * @return The squared distance.
      */
-    fun distanceSq(other: Vector3f): Float {
+    infix fun distanceSq(other: Vector3f): Float {
         val dx = other.x - x
         val dy = other.y - y
         val dz = other.z - z
@@ -271,7 +271,7 @@ data class Vector3f( // @formatter:off
      * @param other The other vector.
      * @return The distance.
      */
-    inline fun distance(other: Vector3f): Float = sqrt(distanceSq(other))
+    inline infix fun distance(other: Vector3f): Float = sqrt(distanceSq(other))
 
     /**
      * Calculates the squared length of this vector.