+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## [Unreleased]
 
+### Added
+
+- Missing `distanceSq` and `distance` functions in `Vector2i`
+- Missing `distanceSq` and `distance` functions in `Vector3i`
+- Missing `distanceSq` and `distance` functions in `Vector4i`
+
+### Changed
+
+- Mark `distanceSq` functions on vector types as `infix`
+- Mark `distance` functions on vector types as `infix`
+
 ## [1.0.0]
 
 ### Added